requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
apted==1.0.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
beautifulsoup4==4.13.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2020.6.20
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==5.3.0
cycler==0.11.0
Cython==3.0.12
decorator==5.1.1
Distance==0.1.3
einops==0.6.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fairscale==0.4.1
fastapi==0.103.2
faster-coco-eval==1.6.7
ffmpy==0.3.2
filelock==3.12.2
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio_client==0.2.6
grpcio==1.62.3
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
hydra-core==1.0.7
install-jdk==1.1.0
isodate==0.6.1
itsdangerous==2.1.2
jieba==0.42.1
lmdb==1.6.2
lxml==5.4.0
matplotlib==3.5.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.6
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.11.0.86
openpyxl==3.1.3
orjson==3.9.7
packaging==24.0
pandas==1.3.5
pathlib==1.0.1
pathlib2==2.3.7.post1
Pillow==9.5.0
plotly==5.18.0
protobuf==3.20.0
prov==2.0.1
psutil==7.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.13.6
pycocotools==2.0.7
pycosat==0.6.3
pydantic==1.10.8
pyDeprecate==0.3.2
pydot==2.0.0
pydub==0.25.1
pyparsing==3.1.4
pyrsistent==0.19.3
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.1
rdflib==6.3.2
regex==2024.4.16
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel-yaml==0.15.80
safetensors==0.4.2
scipy==1.7.3
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.7
simplejson==3.20.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.4.1
starlette==0.27.0
sympy==1.10.1
tenacity==8.2.3
tensorboard==2.6.0
tensorboardX==2.4.1
timm==0.9.2
tokenizers==0.13.3
toolz==0.12.1
torch==1.11.0
torchmetrics==0.7.0
torchvision
traits==6.3.2
transformers==4.29.1
uvicorn==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.15.0
